refactor: event

+++ src/main/java/com/example/demo/api/v1/event/mapper/EventMapper.java
@@ -11,6 +11,6 @@ import java.util.Optional;
 public interface EventMapper {
     Optional<Event> selectEventByEventId(Long eventId);
     List<EventKeyword> selectEventKeywordsByEventId(Long eventId);
-    Long insertEvent(Event event);
+    int insertEvent(Event event);
 
 }
